chore: move mock data so it's not seen as a test
