+++ src/__fixtures__/interpretationsMockData.js
