+++ src/components/Interpretations/InterpretationsProvider/__tests__/hooks.spec.js
@@ -1,6 +1,7 @@
 import { renderHook, act, waitFor } from '@testing-library/react'
 import PropTypes from 'prop-types'
 import React from 'react'
+import * as mockData from '../../../../__fixtures__/interpretationsMockData.js'
 import {
     useInterpretationsList,
     useActiveInterpretation,
@@ -18,7 +19,6 @@ import {
     useInterpretationsManager,
 } from '../hooks.js'
 import { InterpretationsProvider } from '../InterpretationsProvider.js'
-import * as mockData from './mockData.js'
 
 const mockQuery = jest.fn()
 const mockMutate = jest.fn()
